requirements.txt
discord.py
selenium
python-dotenv
requests